requirements.txt
numpy
torch
scipy
chatterbox-tts